UI for recommendation done
