+++ src/main/resources/templates/profile.html
@@ -13,11 +13,41 @@
 
     <div class="row" th:replace="fragments/layout :: amazin-navbar"></div>
 
-    <div class = "row p-4">
-        <div class ="col-sm-4">
+    <div class="row p-4">
+        <div class="col-sm-4">
             <p th:text="'Hello ' + ${user.getFullName()}"/>
         </div>
     </div>
+
+    <h4>Recommended Books: </h4>
+
+    <table class="table table-hover">
+        <thead>
+        <tr>
+            <th scope="col"> Title</th>
+            <th scope="col"> ISBN</th>
+            <th scope="col"> Author</th>
+            <th scope="col"> Publisher</th>
+            <th scope="col"> Inventory</th>
+        </tr>
+        </thead>
+        <tbody>
+<!--        <tr th:if="${recBooks.empty}">-->
+<!--            <td colspan="6"> No Recommendations Available.</td>-->
+<!--        </tr>-->
+        <tr th:each="book : ${recBooks}">
+            <td><span th:text="${book.title}"> Title </span></td>
+            <td><span th:text="${book.isbn}"> ISBN </span></td>
+            <td><span th:text="${book.author}"> Author </span></td>
+            <td><span th:text="${book.publisher}"> Publisher </span></td>
+            <td><span th:text="${book.inventory}"> Inventory </span></td>
+            <td><span><a th:href="'/viewBook/'+${book.isbn}">View it</a></span></td>
+            <td><span th:unless="${book.isOutOfStock()}"><a th:data-isbn="${book.isbn}"
+                                                            th:onclick="addToCart(this.getAttribute('data-isbn'));"
+                                                            href="#">Add To Cart</a></span></td>
+        </tr>
+        </tbody>
+    </table>
 
     <div class="row p-4">
         <div class="col-sm-6">
@@ -27,11 +57,8 @@
         </div>
     </div>
 </div>
-
-
-
+<script type="text/javascript" th:src="@{cart.js}"></script>
 <th:block th:insert="fragments/layout :: bootstrap-scripts"></th:block>
-
 
 </body>
 </html>
